Remove this useless assignment to local variable "tipo".
Remove the declaration of the unused 'tipo' variable.
Remove this useless assignment to local variable "pagina".
Remove the declaration of the unused 'pagina' variable.
Remove this useless assignment to local variable "descripcionCategoria".
Remove the declaration of the unused 'descripcionCategoria' variable.
Remove this useless assignment to local variable "descripcionMarca".
Remove the declaration of the unused 'descripcionMarca' variable.
Remove this useless assignment to local variable "descripcionEstrategia".
Remove the declaration of the unused 'descripcionEstrategia' variable.
Rename "add" as this name is already used in declaration at line 83.
Remove this useless assignment to local variable "tipo".
Remove the declaration of the unused 'tipo' variable.
Remove this useless assignment to local variable "pagina".
Remove the declaration of the unused 'pagina' variable.
Remove this useless assignment to local variable "descripcionCategoria".
Remove the declaration of the unused 'descripcionCategoria' variable.
Remove this useless assignment to local variable "descripcionMarca".
Remove the declaration of the unused 'descripcionMarca' variable.
Remove this useless assignment to local variable "descripcionEstrategia".
Remove the declaration of the unused 'descripcionEstrategia' variable.
Rename "add" as this name is already used in declaration at line 182.
Add the "let", "const" or "var" keyword to this declaration of "upSellingMontoMetaPromise" to make it explicit.
Rename "nivel" as this name is already used in declaration at line 737.
Rename "msg1" as this name is already used in declaration at line 801.
Rename "montolimite" as this name is already used in declaration at line 707.
Rename "msj" as this name is already used in declaration at line 727.
Add the "let", "const" or "var" keyword to this declaration of "i" to make it explicit.
Either remove or fill this block of code.
Rename "impresion" as this name is already used in declaration at line 1607.
Rename "item" as this name is already used in declaration at line 1608.
Rename "precio_valorizado" as this name is already used in declaration at line 1760.
Rename "precio_valorizado_string" as this name is already used in declaration at line 1761.
Rename "precio_catalogo_string" as this name is already used in declaration at line 1762.
Rename "precio_valorizado" as this name is already used in declaration at line 1760.
Rename "precio_catalogo_string" as this name is already used in declaration at line 1762.
Add the "let", "const" or "var" keyword to this declaration of "origen" to make it explicit.