Add the "let", "const" or "var" keyword to this declaration of "_idPed" to make it explicit.
Remove this useless assignment to local variable "tipo".
Remove this useless assignment to local variable "pagina".
Remove the declaration of the unused 'pagina' variable.
Remove this useless assignment to local variable "descripcionCategoria".
Either remove or fill this block of code.
Either remove or fill this block of code.
Rename "posicionEstrategia" as this name is already used in declaration at line 1442.
Rename "recomendado" as this name is already used in declaration at line 1443.
Rename "arraySugerido" as this name is already used in declaration at line 1444.
Rename "impresionSugerido" as this name is already used in declaration at line 1446.
Either remove or fill this block of code.
Rename "tipoMensaje" as this name is already used in declaration at line 2116.
Rename "CantidadAnti" as this name is already used in declaration at line 2793.
Rename "PROL" as this name is already used in declaration at line 2821.
Rename "CliID" as this name is already used in declaration at line 2833.
Rename "CliDes" as this name is already used in declaration at line 2834.
Rename "Cantidad" as this name is already used in declaration at line 2835.
Rename "CantidadAnti" as this name is already used in declaration at line 2793.
Rename "DesProd" as this name is already used in declaration at line 2837.
Rename "Flag" as this name is already used in declaration at line 2839.
Rename "StockNuevo" as this name is already used in declaration at line 2840.
Rename "PrecioUnidad" as this name is already used in declaration at line 2819.
Rename "CantidadAnti" as this name is already used in declaration at line 2793.