Remove this useless assignment to local variable "tipo".
Remove this useless assignment to local variable "pagina".
Remove this useless assignment to local variable "descripcionCategoria".
Remove this useless assignment to local variable "descripcionMarca".
Remove this useless assignment to local variable "descripcionEstrategia".
Either remove or fill this block of code.
Either remove or fill this block of code.
Remove the declaration of the unused 'tipoOferta' variable.
Rename "$divMensaje" as this name is already used in declaration at line 924.