Add the "let", "const" or "var" keyword to this declaration of "selectedText" to make it explicit.
Rename "posicionEstrategia" as this name is already used in declaration at line 894.
Rename "recomendado" as this name is already used in declaration at line 895.
Rename "arrayEstrategia" as this name is already used in declaration at line 896.
Rename "impresionRecomendado" as this name is already used in declaration at line 898.
Rename "ValidateOnlyNums" as this name is already used in declaration at line 1761.
Rename "container" as this name is already used in declaration at line 2626.
Rename "txtSaludoIntriga" as this name is already used in declaration at line 2628.
Rename "nombre" as this name is already used in declaration at line 2750.
Rename "tema" as this name is already used in declaration at line 2751.
Either remove or fill this block of code.
Either remove or fill this block of code.
Either remove or fill this block of code.