Remove the declaration of the unused 'ventanaChat' variable.
Remove this useless assignment to local variable "urlBienvenida".
Rename "URLactual" as this name is already used in declaration at line 176.
Rename "urlBienvenida" as this name is already used in declaration at line 177.
Either remove or fill this block of code.
Either remove or fill this block of code.
Rename "params" as this name is already used in declaration at line 498.
Add the "let", "const" or "var" keyword to this declaration of "expr" to make it explicit.
Add the "let", "const" or "var" keyword to this declaration of "montoWebAcumulado" to make it explicit.
Rename "j" as this name is already used in declaration at line 933.