Remove the declaration of the unused 'pos' variable.
Either remove or fill this block of code.
Remove the declaration of the unused 'descripcion' variable.
Remove the declaration of the unused 'ItemTotal' variable.
Remove the declaration of the unused 'hdfTotal' variable.
Remove the declaration of the unused 'CliID' variable.
Remove the declaration of the unused 'DesProd' variable.
Remove the declaration of the unused 'CliDesVal' variable.
Remove the declaration of the unused 'CantidadAnti' variable.
Remove the declaration of the unused 'ClienteAnti' variable.