Rename "obj" as this name is already used in declaration at line 119.
Remove this useless assignment to local variable "urlAccion".
Rename "Cantidad" as this name is already used in declaration at line 357.
Rename "obj" as this name is already used in declaration at line 404.
Remove this useless assignment to local variable "descripcionMarca".
Rename "Cantidad" as this name is already used in declaration at line 647.
Rename "j" as this name is already used in declaration at line 812.
Remove this useless assignment to local variable "mensajePedidoCheckout".
Rename "tipoMensaje" as this name is already used in declaration at line 868.