Either remove or fill this block of code.
Remove the declaration of the unused 'revisarMostrarContenedorCupon' variable.
Remove the declaration of the unused 'enviarCorreoGanasteCuponPromise' variable.
Remove this useless assignment to local variable "mostrarContenedorConocelo".
Rename "backImg" as this name is already used in declaration at line 613.
Rename "nuevoBackImg" as this name is already used in declaration at line 614.
Rename "backImg" as this name is already used in declaration at line 626.
Rename "nuevoBackImg" as this name is already used in declaration at line 627.