Rename "obj" as this name is already used in declaration at line 113.
Remove the declaration of the unused 'urlAccion' variable.
Rename "Cantidad" as this name is already used in declaration at line 349.
Rename "obj" as this name is already used in declaration at line 413.
Remove the declaration of the unused 'descripcionMarca' variable.
Rename "Cantidad" as this name is already used in declaration at line 665.
Rename "j" as this name is already used in declaration at line 829.
Remove the declaration of the unused 'mensajePedidoCheckout' variable.
Rename "tipoMensaje" as this name is already used in declaration at line 953.